+++ packages/swagger2apis/src/plugins/Compatible.ts
@@ -1,16 +1,11 @@
 // 兼容以前版本的一些插件
 import type { IPlugin } from "../plugin";
 import { firstUpperCase } from "@pdcode/utils";
-import { printErrInfo, printWarnInfo } from "@pdcode/utils";
+import { printWarnInfo } from "@pdcode/utils";
 
 // 兼容以前版本的请求方法名
 export const CompatibleApiFnNameOnOldversionHandle: IPlugin = {
   befofeRender(ctx) {
-    printErrInfo(
-      "[CompatibleApiFnNameOnOldversionHandle]老版本方法名兼容插件开始工作\n",
-      "因为以前的命名方式某些情况下会导致重复key的出现,所以新版采用了新的命名方式来避免,当发现有重复key的出现,后续重复的key将不再执行兼容处理~\n",
-      "该插件会尝试尽可能的保留三个版本的方法名:以前版本的byCode和byCode[codename],以及新版本的$[codeName]$\n"
-    );
     const { renderData } = ctx;
     const { apis } = renderData;
 
@@ -52,6 +47,6 @@ export const CompatibleApiFnNameOnOldversionHandle: IPlugin = {
         return api;
       })
       .flat();
-    printWarnInfo("[CompatibleApiFnNameOnOldversionHandle]插件处理完成,但无法保证覆盖所有case,若需要可自行实现插件覆盖需求~");
+    return ctx;
   }
 };