+++ packages/swagger2apis/src/plugins/Compile.ts
@@ -1,17 +1,18 @@
 import { compileTs, writeFile } from "@pdcode/utils";
-import { join } from "path";
 import prettier from "prettier";
 import type { IPlugin } from "../plugin";
-import { printInfo, printSuccInfo } from "@pdcode/utils";
 
 export const createCompileTS2JSPlugin = (tsconfig = {}, prettierConfig = {}): IPlugin => {
-  const blackExtNameList = ["d.ts"];
+  // 下面的文件不进行编译
+  const blackExtNameList = [".d.ts"];
   return {
     afterWriteFile: async (ctx) => {
-      printInfo("开始编译TypeScript文件~");
       const { writedFileList } = ctx;
-      for await (const filePath of writedFileList) {
-        if (blackExtNameList.every((extName) => filePath.includes(extName))) continue;
+      // 过滤出ts文件
+      const tsFiles = writedFileList.filter((path) => {
+        return path.endsWith(".ts") && !blackExtNameList.every((extName) => path.endsWith(extName));
+      });
+      for await (const filePath of tsFiles) {
         const res = await compileTs(filePath, tsconfig);
         const content = await prettier.format(res.outputText, {
           parser: "typescript",
@@ -20,11 +21,13 @@ export const createCompileTS2JSPlugin = (tsconfig = {}, prettierConfig = {}): IP
           singleQuote: false,
           ...prettierConfig
         });
-        const filePathNoExtName = filePath.split(".").at(-2);
+        const jsFilePathNoExtName = filePath.split(".").at(-2);
+        const jsFileFullPath = jsFilePathNoExtName + `.js`;
+        writedFileList.push(jsFileFullPath);
         // 将编译后的 JavaScript 文件写入磁盘
-        writeFile(join(filePathNoExtName!) + `.js`, content);
+        writeFile(jsFileFullPath!, content);
       }
-      printSuccInfo("[createCompileTS2JSPlugin]TypeScript编译插件已经完成对TypeScript文件的编译~");
+      return ctx;
     }
   };
 };