+++ packages/swagger2apis/src/plugins/PrefixRemove.ts
@@ -1,7 +1,6 @@
 // 擦除url前缀插件
 
 import { IPlugin } from "../plugin";
-import { printSuccInfo } from "@pdcode/utils";
 
 export const createPrefixRemovePlugin = (prefix: string): IPlugin => {
   return {
@@ -13,7 +12,6 @@ export const createPrefixRemovePlugin = (prefix: string): IPlugin => {
         newPaths[(path as string).slice(prefix.length)] = paths[path];
       });
       ctx.rawJSON.paths = newPaths;
-      printSuccInfo("[createPrefixRemovePlugin]已经完成对url前缀的擦除~");
       return ctx;
     }
   };